LATTICE-2102 move allowed entity type checks to helper class
